Route CLI - Silently ignore errors regarding vlan metadata
Summary:
Route CLI fetch the metadata about ports and vlan from Agent/Qsfp services.
However the part of the information is not available in Emulation. With this
diff we make fetching of this information optional so that users can leverage
FBOSS CLI on emulated nodes and inspect routing state.

Minor Change in RIB - `Route::toRouteDetails` doesn't populate the new field.
This part of code is only exercised in emulation hence production is not affected.

Reviewed By: jasmeetbagga

Differential Revision: D26504141

fbshipit-source-id: 68a05fd5aaac04852a1183a63173c72761309176