reload object without deleting temporary
Summary:
reload object must be idempotent. that is it should not lead to deleting sai object in SDK (i.e. triggering SAI API's remove method to remove object).

however there was a bug in SaiStore reload object.

a temporary object is constructed before store attempts to ref or inserts it. if object already exists, ref map returns existing object which then triggers destructor of temporary and attempt is made to remove sai object in sdk.

if sai object exist, specifically mark that object must not be specifically removed.

Reviewed By: jasmeetbagga

Differential Revision: D26649080

fbshipit-source-id: 7baeecceff09a1b80c57477aa057440a1787acde