Standalone module fw upgrade: Multithread support for firmware upgrade on multiple modules in different threads
Summary:
This checkin supports the firmware download on multiple modules in a different thread. We have found that if two modules on the same I2c controller are upgraded simultaneously then we get many i2c errors so we will do parellalization in such a way that the modules belonging to different controllers will be upgraded at the same time in context of different threads.

The CLI will take the range of ports for doing the firmware upgrade. It will first check which all ports are eligible for upgrades:

1. The modules which are present
2. Modules which are of CMIS type
3. Modules which matches type provided in CLI (ie: finisar-200g)
4. Modules which does not have the same version as being upgraded

This is first list of modules. After this depending on how many ports the i2c controller serves in the platform, the list of modules will be divided in different bucket. This process is called bucketization. The modules will be divided in buckers such as one bucket contains the modules belonging to the same controller. Each bucket will be served by a separate thread. This way we will not have two port upgrade happening simultaneously on the same controller. Example:

Lets say we have to upgrade the modules 2,3,8,16,33,38. Let’s say this is Yamp platform and each controller controls 8 modules. The we will divide these ports in 3 buckets:

Bucket1: 2,3,8
Bucket2: 16
Bucker3: 33,38

We will create a thread for each bucket. So the firmware upgrade will run on modules 2,16,33 in parallel. But the upgrade on 3 will wait for 2 to finish in same thread. This way there will not be any conflict at controller level.

CLI:

./wedge_qsfp_util  <portA> <portB> —module_bulk_fw_upgrade —firmware_filename <filename> --module_type <module-type-str> —fw_version <version-str>

Here portA and portB specifies the range of port on which the firmware upgrade has to be done. The upgrade will find the list of ports between them as specified above

Filename is the name (with path) for the firmware

Module-type-str is string specifying the module type like: finisar-200g, innolight-200g

Version-str is the version string which is two hex bytes separated by a dot, example: 7.8

Example of the CLI

[root@yamp tmp]# ./wedge_qsfp_util 1 100 --update_bulk_module_fw --firmware_filename /var/facebook/fboss/FW_36965_FTCC1112E1PLL-FB_7p8.bin --module_type finisar-200g --fw_version 5.6

Reviewed By: birdsoup

Differential Revision: D24107452

fbshipit-source-id: 6b07333c043ec0464997ec8ed3313ed702b87cbe