Standalone module fw upgrade: Added port regex support
Summary: Added port list simple regex support. The list of modules to be upgraded can be specified using —port_range option and its value will be a string in format of: 1,3,5-8. Example: the input 6,9,15-18,33 will pick up the ports 6,9,15,16,17,18,33 and check which all optics are qualified for upgrade. Then it will create a final list of optics to upgrade based on module type, current firmware version. After that this list will be bucket’ized and then assigned to those many threads to upgrade in parallel.

Reviewed By: birdsoup

Differential Revision: D24236922

fbshipit-source-id: 961e06d18a684dda428aed4bafa1338a634a3306