Adding more fields to TransceiverInfo
Summary:
This is the first change in a series of changes required for moving all the DOM parsing logic from wedge_qsfp_util into qsfp_service. The change in this diff adds some fields to TransceiverInfo thrift structure which will be used in future diffs by the wedge_qsfp_util client.

It's possible that certain fields don't have the correct or updated value right now. The next diff will add a unit test for this change, and fix any problems that are found after the test.

Differential Revision: D26239031

fbshipit-source-id: 43f35452aff24b96913f6298106b6e4bd8fe7bce