remove unnecessary check
Summary:
as titled. making it simpler. this check requires that port handle must be placed before creating vlan member. this is not required because vlan member is really a subscriber which gets active once bridge port is ready.

config validation is best delegated to SwSwitch and not to HwSwitch or SaiSwitch.

the check doesn't make sense,  invalid config gets rejected in swswitch

Reviewed By: jasmeetbagga

Differential Revision: D26656139

fbshipit-source-id: 3ac2f440ff9602500deda1d735fae042fbae93e3