Make wedge_qsfp_util cli work for blanco bypass module
Summary: The Blanco bypass module has the module id as CMIS but it does not have any data so the code does not contain the cache for this. Due to this when refresh() is called from qsfp_service then it throws. We need to catch the exception in refresh() when called from wedge_qsfp_util. The exception is already caught in refresh() call inside the qsfp_service.

Differential Revision: D26643264

fbshipit-source-id: c4c96f0ba12b958aaead2b29364984f22620b437