Check qsfp_service running before invoking firmware upgrade
Summary: When the qsfp_service is running then the firmware upgrade CLI may fail. The f/w upgrade does thousands of i2c operations and if qsfp_service is running then it will fail due to i2c error. If the user is trying to do firmware upgrade and qsfp_service is running then warn the user to stop qsfp_service and try the command again.

Differential Revision: D26643713

fbshipit-source-id: c8e2375f63bb2c80002182eb02e3438f817b9b81