drop severity of missing transceiver log
Summary:
This drops the severity of this log. A lot of times, it is
actually expected that we don't have a transceiver in cache, so
logging each time is unnecessary.

We should improve QsfpCache so it is easier to distinguish expected
port missing from errors communicating with service, but for now let's
reduce log noise.

Differential Revision: D26838670

fbshipit-source-id: 69fae548e43cdc4ef7948d373773c7509abd6d16