vlan member adapter host key
Summary:
while working on default vlan, i realized a bridge port always belong to default vlan even if it is part of another vlan.

in general, a port or bridge port may belong to multiple vlans. this is manifested with default vlan in our environments because in our environment every port is always part of one configured vlan. however sdk's default vlan is not configured by fboss and still bridge port always belongs to that default vlan.

to handle this, modifying adapter host key of vlan member to ensure that it is identified by both vlan and bridge port, and not just bridge port.

not doing this sort of muddles the way object get claimed. it may so happen vlan member of default vlan may get stitched to another vlan and vice versa because adapter host key of vlan member doesn't identify fully the vlan member.

Reviewed By: jasmeetbagga

Differential Revision: D26756091

fbshipit-source-id: bdcbe46194c5cfb36a39f11b7a02d341b4aae0b3