Begin coding rollback algorithm in RIB
Summary:
If HW rejects a update, the switch state is adjusted to
reflect what got into HW. For correctness we need to have the
RIB match the FIB as well. The simplest way to achieve this
is to rollback the entire route update from FIB and RIB.
We will then return the rolled back Hw switch state in
FbossHwUpdate error.

Start coding up the simplest case, we simply rollback the
routes we added, deleted. So deleted routes are added back
and added routes are deleted.

Many TODOs here. This only addresses the simplest case of
non overlapping adds and deletes. Subsequent diffs will address
these TODOs one by one

Reviewed By: phshaikh

Differential Revision: D26679465

fbshipit-source-id: bfa6c350b0c04b905c75e354f74087f7fe98955c