Have RouteUpdater::addRouter return route it replaces (if any)
Summary:
This is essential bookkeeping for accomplishing rollback
and is similar to what we did for del routes in previous diffs.

Note that this does add some memory for bookkeeping but its no
worse than what was in legacy rib, since there the old routes
(from all clients) were maintained in old switch state, which
lived for duration of delta processing. Maintaining replaced,
deleted route entries in new RIB is equivalent.

Reviewed By: phshaikh

Differential Revision: D26702582

fbshipit-source-id: 16dc30b6f969a7a8d1b18e9f492ee0fc28362c3e