Process CREAETE_ONLY speed attribute.
Summary:
The SAI specification does not mark SAI_PORT_ATTR_SPEED as CREATE_ONLY.
However, that is a requirement for certain ASICs.
Previous stacked diff introduced HwAsic Feature for that. Consume it.

Working with Broadcom if there would be a better method to
implement this (e.g.
a SAI switch attribute), but in the meanwhile, unblock with
this.

Differential Revision: D26899393

fbshipit-source-id: 31a21b5c941774708e4c1a424ee7d27b52a5e416