Feature for ASICs that have Port SPEED as CREATE_ONLY attribute
Summary:
The SAI specification does not mark SAI_PORT_ATTR_SPEED as CREATE_ONLY.
However, that is a requirement for certain ASICs.
Introduce HwAsic Feature for that.

Working with Broadcom if there would be a better method to implement this (e.g.
a SAI switch attribute), but in the meanwhile, unblock with this.

Reviewed By: jasmeetbagga

Differential Revision: D26899395

fbshipit-source-id: cf957dd07690d62557eb2516115c37ce32031d16