Create HwLoopBackTests to verify looped back packets are not dropped
Summary:
In some cases due to brcm-sai implementation, looped back packets would be dropped because of the srcMac == dstMac. This tests is created to ensure that looped back packets are working properly.
The current brcm-sai implementation does drop the packet when it's originated from CPU port. Will need to file CSP for brcm-sai after this test is landed.

Reviewed By: shri-khare

Differential Revision: D26864527

fbshipit-source-id: 3aad8b4168e6c956522639dd7e41fcb14982b1ae