fix test failures due to unchecked access to optional
Summary:
qsfp_service tests are failing due to recent diff that added
logic to ensure rx squelch is enabled. This fixes the logic to only
enter that logic if we have the hostlanesettings populated, which is
not the case in tests.

Currently this shows up in fboss/wedge_agent contbuild, but it should
be split out (see D27034650)

Reviewed By: rajank7

Differential Revision: D27034718

fbshipit-source-id: 7f686429456d9da762bf93404a7d964c6e2fe3fc