Add a delLinkLocalRoutes API to rib RouteUpdater
Summary:
When all interface routes disappear for a VRF,
we prune link local route added for that VRF as well.
This is the existing in legacy RIB, mimic that in new RIB.
This doesn't really matter in practice since running w/o
any L3 intfs configured for a VRF is not something we do.
But it matters for tests so mirror the behavior

Differential Revision: D26942457

fbshipit-source-id: a0d4547ab40f58cdb73976d097fbd0e0cedb114d