in hardware learning ignore unclaimed dynamic fdb entries
Summary:
as titled.

if learning is hardware, its possible some mac is learned and remain unclaimed because warmboot switch state may not have corresponding entry and remain unclaimed.

if learning is software, expect all entries in switch state and so all entries in warm boot be claimed.

added a method to inject closure for counting unclaimed objects. return true if entry should be counted, else return false. entire sai object is available in closure so any property of object can be used to determine if entry can be counted or not.

Reviewed By: jasmeetbagga

Differential Revision: D26761253

fbshipit-source-id: 4eff65408d1fc2f5214367b7d568068ca0cce396