Get rid of hacky COPY_PREFIX_AND_NEXT_HOPS behavior
Summary:
What we need during route resolution is to copy the old
route, but clear its forwarding info and start reresolution. We did
this by adding another means of copy instead of just using the
copy constructor. We were careful to copy the no fwd info fields
of the day. This broke when routes got class IDs. This may have
lead to us not comparing class Ids in route == operator as well,
since we weren't copying them here. Confirming the theory that
one hack leads to another. Get rid of the original sin.

Reviewed By: tfangit, aeckert

Differential Revision: D27178068

fbshipit-source-id: 330617e75c757c53e1c07764da4241d5d8afecd8