Add unimplemented API support for Bcm Cinter
Summary:
Add Bcm Cinter support for some of the unimplemented APIs. Although most of them are not used in wedge agent, we want to add them such that no runtime error is introduced by BcmCinter.

Some of the remaining APIs contain complicated bcm struct and will be handle in later diffs.

Reviewed By: shri-khare

Differential Revision: D27096078

fbshipit-source-id: f72581d8988508e2e767ab6846063e5d007bfa12