Allowing firmware upgrade for corrupted optics
Summary:
We found that when firmware upgrade is interrupted in middle on Innologht 400G optics then it does not contain a valid firmware. After that all its register values show up as 0 (except part no).

In our firmware upgrade CLI there is a check to see if it is CMIS module. For such corrupted module we need to allow non identified module also after printing a message.

Differential Revision: D27139151

fbshipit-source-id: 5667c9d81a2b4ac994f53cce54b97484a6c4edd3