When computing CIDR n/w apply mask
Summary:
As tittled. Sometimes in tests we set more than
the mask bits on RoutePrefix objects, for converting to Cidr network
get significant bits only. This distinction was leading to
a subtle bug in BcmSwitch, where we would try to find
and revert overflown routes in switch state. However routes
in switch state (or in HW) have only significant bits set,
so take care of this in the API.

Differential Revision: D27225489

fbshipit-source-id: d53f6c988bd9abedb58f5aa573278016d4141a76