Start introducing shadow ribs for lookup
Summary:
Add shadow route tables for lookup. During route updates and later
resolution, we make updates to RouteTables and then trigger a
reresolution. During this stage the RouteTable is not in
a state where we can use it for lookup. E.g. any of the new
routes added don't have any forwarding info (which comes after
resolution) so we can't use this data structure for both
resolution and lookups. Hence a shadow route table to
facilitate lookups. Regular and shadow route table will
share routes via pointers, but shadow tables will be updated
only after resolution and will always present a sane state for
lookup.

While this is extra memory, the cost is not prohibitive. At 100K
routes (say) the memory cost in 100K shared pointers, which tranlates
to 1.6Mb of memory, which is pretty tolerable.

TODO:
Future diffs will ensure that these two sets of tables are kept in
sync and will add optimizations

Differential Revision: D27205555

fbshipit-source-id: f63ad3805be59cbdb703a96a4bb5a2c84675fb3d