On HwUpdateFailure rollback rib using appliedState fib
Summary:
With the sharing of Route pointers fib has the same information
as RIB and can be used for rollbacks. Do that instead of using shadow rib.
Two reasons for this
1) (minor) More efficient in terms of lookups since we don't have to
hold the shadow rib lock
2) (major) Shadow rib needs to be update before we send updates to
FIB (i.e. immediately post resolution. Reason for this is that
the observers are free to do longest match lookups. And since we
plan to do longest match lookups via shadow rib, we need to match
it what we are sending to HW. Next diff will make that change

Differential Revision: D27247389

fbshipit-source-id: 80133abf11f6a9b82331f963ba61d97d6df9f194