Dont change speed on SAI TH3
Summary:
We discovered a bug in BRCM-SAI implementation that causes ports to disappear
on speed change. e.g. port 1 can change speed from 100G to 200G, but attempt to
do the same on port 5, causes the port to disappear (the lanes go missing as
well).

Broadcom is investigating this as part of the referenced CSP.

Hard code max speed to 100G on all ports for now to unblock testing.

FWIW, native BCM get max speed implementation always returns 100G, that is a
bug we need to fix anyway, but with a fix such as this, SAI tests would run
similar to how native BCM tests run today.

Differential Revision: D27342837

fbshipit-source-id: 684a4dde4cdfd06d1897868e5ff60cd1ae844dc6