Fix QSFP Presence detection on Wedge100
Summary:
It turns out that the QSFP scanPresence logic on Wedge100 is buggy. The presence bits are supposed to be swapped between adjacent modules. Bit0 indicates presence for QSFP1, bit1 for QSFP0, bit2 for QSFP3, bit3 for QSFP2 and so on. The current logic assumes its a 1:1 mapping.

This is not causing any problems right now because we just care about the number of present transceivers and the count will still be correct with the bug.

Reviewed By: rajank7

Differential Revision: D27340512

fbshipit-source-id: cd46bedd2dde310fc8f06093900325cbb4110b49