Build coarser route update interfaces using bulk route updates
Summary:
Experimenting with fine grained locking, whereby
we can hold the lock just for the rib route update and reresolution
interval and not spray it across thrift conversions, individual
route adds etc. To do that we would need to leverage a coarser
route update interface.
As a experiment, I clocked the rib udate with 30K routes with
a lock duration of 500ms, with optimizations we may be able to
bring it to a acceptable level for route lookups

Differential Revision: D27283865

fbshipit-source-id: bbd0da547245dca7bb8a4e0bb9a5081bc6d789d4