Bye bye shadow rib
Summary:
With the optimizations in previous diffs, we are down to a
the following worst case lookups times
- Worst case single lookup competing with route prog: 220ms
- Worst case 1000 lookups competing with route prog: 225ms

The route scale benchmark from previous diffs, makes route lookups from
a separate thread compete with the following
- Route add with 4k route chunks
- SyncFib with full route scale
- SyncFib with full route scale but all nhops changed to width N + 1
- Delete all routes (full route scale)
- Readd all routes (full route scale)

The above was benchmarked with FSW scale (16K routes) on TH atom and
HGRID UU scale (50k routes) on YAMP. w400/400c/elbert/fuji all have
the same CPU as YAMP so we expect similar results.

With this we have a 220ms worst case for a single lookup, that's the
worst to expect in a (say) packet rx handling lookup path. Its important
to note that while a single packet may see this delay, back to back
packets cannot (hence the numbers for worst case 1000 lookups). This
is because  wlock is only held during the rib update + resolution
period, after that we do HW writes which is a multisec interval
during which rib can be freely looked up from. Writes are serialized
so we don't have to worry about wlock contention. This seems a
acceptable interval, considering that we aren't a real time OS and
could well have suffered a 200ms thread scheduling delay.

We still have more optimizations in our pocket if we choose so
- Break the route add/del to 4K chunks. This will bring down the
contention interval to approx 120ms
- Share next hop memory - TBD on how much this will help, but seems
it can help in terms of more efficient lookups on nhop resolution
cache, since we can just hash, compare pointers instead of more
expensive next hop set comoparisons

Differential Revision: D27401452

fbshipit-source-id: 4c09f8910acb9706f667d83ba8556754df25677a