Fix build error in //fboss/agent/test:utils
Summary: Our local copy of googletest used to have a patch that caused gmock matchers to take rvalue refs. The new version removes this patch. In order to account for both signatures, we use a preprocessor directive to switch based on whether we have the old or new verison of googletest.

Differential Revision: D27351205

fbshipit-source-id: d3e3b8aa1741a66d9aeea90b2ac79b6dcb820018