Go to a less often watermark stats collection ala native BRCM
Summary:
Quere watermark stats collection is pretty expensive. Ultimately
on brcm this results in a synchronous call to HW followed by a update
to quantile stats. In native BCM world  we default to collecting this
every 60s, but were collecting them every second on SAI. This was
prohibitively expensive. Concretely the TH alpm route scale benchmakr
now runs 40% faster with this change. So we are down from 30s to ~18s
for this benchmark, which is actually pretty close to native BCM SDK
number. Similar wins can be expected in other programming intensive
benchmarks as well.

Differential Revision: D27418311

fbshipit-source-id: cf132f814f606ecab7d8775f31a8a9a1f6efa980