Revert D27306550 and D27348149
Summary:
Not sure why, but these changes seem to cause us to fail to initialize phys on minipack. Backout out the changes for now.

It doesn't seem like we should actually require this dependency if we're not using MDIO via the TH3, but the core reset function seems to fail without even doing any mdio reads/writes when we're missing this dependency

Reviewed By: aeckert

Differential Revision: D27522876

fbshipit-source-id: 61668f5bb84439310649708092462cd4bf73befb