Write new boot header and current time
Summary:
At the start of a new Bcm Cinter / Sai Replayer log, write a new boot header and the current timestamp.

This would be easier for the cint gather tool to determine the start of a cold/warm boot, properly split/merge the compressed log to multiple cint files, and then replay them based on the timeline.

Reviewed By: shri-khare

Differential Revision: D27484363

fbshipit-source-id: 66612fce5bc5533431d9f3fdcd3c647efe24a705