Let qsfp_service crash without agent config
Summary: At provisioning, it is still not guaranteed that the config will be generated before we bring up the process. On Wedge400, using default settings will silently cause module to be not programmed. Currently agent will crash w/o the config and here making qsfp_service doing the same. In future we will migrate this dependency onto qsfp config instead of agent config.

Reviewed By: rajank7

Differential Revision: D27472854

fbshipit-source-id: 27811ed03ed0886af2e31e1a07b4e4fe865b7b2b