Adding support for forcing a cold boot of qsfp service
Summary:
This change adds a way to force a cold reboot for qsfp_service. When the file "/dev/shm/fboss/warm_boot/cold_boot_once_qsfp_service" is touched and then the qsfp_service restarted, the QSFP modules are hard reset which causes a cold boot.

The ultimate goal of this change is to allow cold boot testing in netcastle.

Reviewed By: birdsoup

Differential Revision: D26968575

fbshipit-source-id: 7ed261e3506c4fff64d8c3954d50ed0fe8dac50b