make HwStatsCollection use same number of ports
Summary:
This diff tries to produce consistent stats collection benchmark numbers for minipack
and wedge400 by enabling the same number of ports in the test.

More specifically
change to BcmSwitch::getPortStats(), fixed a potential performance issue when it is
called for disabled port that was never programmed, i.e. programmedSetting_ is null.
With out this fix, a lot of time is wasted in BcmPort::isEnabledInSDK(), see P360911805.

change to enableLinkscan() in BcmPort.cpp, avoid the impact of linkscan on disabled
ports. Currently, linkscan is always enabled on all ports, including disabled ports.
As a result, linkscan thread on minipack is always busier than wedge400, even
with the same number of ports enabled. It seems stats collection thread performance
would be affected by linkscan thread due to underlying SDK locking mechanism. So,
this diff enableLinkscan() only when the port is already enabled or programmed for the
first time.

Reviewed By: jasmeetbagga

Differential Revision: D27454466

fbshipit-source-id: 79b0038eeb6a4840bba7aa396cc6e43493cd35f0