Fix for cpuPort_ in saiSwitchManager to avoid object busy issue
Summary:
After the checkin D27401834 (https://github.com/facebook/fboss/commit/f12d21af3d292559799dbc03afca43b958883fbd), the Credo Phy is not coming up. The Sai bring up for Credo Phy fails with this error:

what():  [switch] Failed to remove sai object : SwitchSaiId(1): OBJECT IN USE

The Phy does not need the CPU port so the CPU port init is being moved to a different file and for Phy it will be blank function. The getCpuPort will return cpuPort_ for npc but for Phy it will throw exception

Reviewed By: phshaikh

Differential Revision: D27538994

fbshipit-source-id: 17a4992e6392151c61ca598cf3eab4ccc3a41856