Serialize full rib till optimization to recover rib from fib is in prod
Summary:
As titled, I added a optimization to reconstruct RIB from FIB.
That allows us to optimize how much of the RIB we need to serialize.
However, for warm boots from trunk -> prod to work the reconstruction
code needs to get to prod wedge_agent. Till then continue to serialize
the entire RIB

Differential Revision: D27560096

fbshipit-source-id: a73bdfd80c1ff583b32800dac645ad511dcb3795