Create python script to modify Sai log and compile using gcc
Summary:
Currently we only build Sai Replayer binary through buck or OSS build, and this could be inaccessible for vendors such as Sai team to compile.

This python script will modify the Sai log to a standalone file, and then compile ot isomg gcc with the provided libraries.

(qq: do we need a buck target for this?)

Reviewed By: shri-khare

Differential Revision: D27555438

fbshipit-source-id: ce5904f651fe44ba461485464604eb63e27a6a97