Improve getSafeProfileID() logic
Summary:
In this diff, I want to deprecate the current getSafeProfileID() logic in ConfigFactory.
The current logic is to find the profile of a port which doesn't need to subsume other ports.
It might be wrong that if we mixing using different "safe" profile in the same group.
For example, like 10G and 25G profiles both don't use any subsumed port,
so if ports in the same port group mix using 10G and 25G profiles, then hardware won't allow to program in such way.

Therefore, I move the logic of finding best(safe) profile for ports in the same group out from previous securePortsInConfig()
to a new helper function getSafeProfileIDs().
Now both setPortToDefaultProfileIDMap() and securePortsInConfig() can use getSafeProfileIDs() to find the safe profile for all ports in the same group.

Differential Revision: D27476534

fbshipit-source-id: d7c42dc578f1a9c7146f363110bebe4b92ebdfd6