Remove same profileID check in config generation
Summary:
Because we added securePortsInConfig() in D27445544,
which should be able to make sure the ports in the same group to use the profile that won't conflict.

Now, we can remove this extra check to make sure all ports from the same core to use the same profile id.

This will make the whole config generation logic a little bit simpler and faster.

Differential Revision: D27474642

fbshipit-source-id: 00fc3337cc4f3b5bb51efadfb55bc0613385931e