Introduce securePortsInConfig in ConfigFactory
Summary:
This diff introduces securePortsInConfig().

This function is to secure all ports in the input `ports` vector will be
in the config. Usually there're two main cases:
1) all the ports in ports vector are from different group, so we don't need
to worry about how to deal w/ the slave ports.
2) some ports in the ports vector are in the same group. In this case, we
need to make sure these ports will be in the config, and what's the most
important, we also need to make sure these ports using a safe PortProfileID

This is intermediate diff that before we switch to use the static PortToDefaultProfileIDMap
to generate the port config. Because PortToDefaultProfileIDMap is using the SwState portMap at the initialized stage,
the map might not have all the ports, especially slave ports.
Using this securePortsInConfig() to make sure we can have a valid port config in the next diff.

Differential Revision: D27445544

fbshipit-source-id: 37296893dfe3d9fa929a8a7b9a96fc6d78ab88fa