Enable fixed width wide ecmp normalization in normalizedNextHops() call
Summary: Normalization of wide ecmp to a fixed width size was handled in hw layer since this is specific to TH3. However this has a side effect that show commands will not display normalized weights. I am adding this to normalizedNextHops() call so that correct weights are displayed. Also added a flag to control this. We still need the logic in hardware layer to take care of ecmp shrink/expand scenarios.

Reviewed By: jasmeetbagga

Differential Revision: D27435257

fbshipit-source-id: f05e6d9f8f02a9a78f820bf34d782e765da5afb6