Change the wide ECMP normalization algorithm
Summary: The existing ECMP normalization algorithm had the goal of preserving weight ratio across different weight groups as close as possible to the desired ratio. However this has a side effect that within a weight group, there could be disparity among member links because we try to distribute the extra slots among all weight groups. After discussions with routing and controller team, it appears that a better option is to distribute the extra slots to higher weight nhop group and keep the percent deviation within a group as small as possible.

Differential Revision: D27014407

fbshipit-source-id: ba3d190e5ef4e073b2070299ea7444adbb242491