Wedge100, Galaxy : Detect QSFP presence by reading the status of hardware pin
Summary:
Currently, the QSFP presence on Wedge100 and Galaxy is read by reading the identifier byte 0. If the module is present, the read is successful. If the module is absent, read throws an exception. This makes it difficult to differentiate between the two cases 1) Module is absent 2) Module eeprom read times out due to a different issue.

To avoid this situation, we should instead detect the presence by reading the status of hardware presence pin. The status is exposed in the CP2112 I2C space on Galaxy and Wedge 100. This will allow presence to take on a tristate value present, absent, unknown in future.
Other platforms like Minipack, Yamp etc already detect the QSFP presence by reading the status of hardware pin from the FPGAs.

Differential Revision: D27249627

fbshipit-source-id: e436a5464cc28fddb2009cae8c4a3cfa43a1f267