Program pfc maps only if there is a change.
Summary:
Ensure that we program the PFC maps when there is a change and not to program during warmboot. This can cause issues in MMU if traffic is running. So do get() and do write() only if there is a change.

Refactored read of pfcs maps to QosPolicy and reuse across the tests.

Differential Revision: D27496336

fbshipit-source-id: 075f8a215707b331169b1d332802d32aaa178d3b