use portOffset instead of portIndex in galaxy led
Summary:
In T88021883, ENS saw lots of galaxy switches w/ no LEDs on. It seems that it's only on SAI galaxy switches, and the non-sai implementation seems to still work on these switches.

It seems that using the portOffset instead of the port index causes the led's to work correctly again

From CS00010070622, it looks like the last guidance from BCM was to use the portOffset when programming the LED state, but it seems our current implementation had been verified in T76556789 (unless we were using a different implementation at the time). I'll follow up w/ a new case with broadcom to see if there's been a regression/ask them what their current stance is on how we need to program the LEDs

Reviewed By: phshaikh

Differential Revision: D27748910

fbshipit-source-id: 62fc97fc3072268f9127dc22da918021b5a668fc