Log warning when multiple nhops have classID
Summary:
Queue-per-host semantics can break if a route has multiple resolved nexthops (with
classIDs). This is because owing to hardware limitation, a Route can only
inherit classID of one of the nexthops but any of the nexthops could be used
to route packets.

More details in the comment in the code.

This diff detects that and logs it to the agent log viz. :
/var/facebook/logs/wedge_agent

Reviewed By: jasmeetbagga

Differential Revision: D27817180

fbshipit-source-id: 5b48fdb6b72a0b5ddfeeaa37336ff973eb63c571