Change Elbert and Yamp pim indices to start from 2
Summary: Change pim start to 2 for Yamp family of platforms. I've wanted to do this for a while but its been hard since this is spread out in a lot of our codebase. I opted for doing it all at once because I think that should minimize the chances of off by one bugs.

Differential Revision: D27682336

fbshipit-source-id: 5586e2b81ce5290552ceaca5fa9663a5470b93b5