Pass the boot type information from wb helper to async logger
Summary: To precisely determine when the coldboot/warmboot happen, as well as where to start in the cint/sai replayer log file, warmboot helper is modified to pass the boot type to async logger. Since async logger might not be initialized during warmboot helper is determining the boot type, it would pass the information through a static function.

Reviewed By: jasmeetbagga

Differential Revision: D27896260

fbshipit-source-id: 6e7dd53de1643ad0d26aaeb75f3e7dcad02235d0