Fix Copp tests on TH4
Summary: All BcmAclCoppTests were failing because of an sdk bug. The issue is that when we create an ACL with queue id 0, reading back the value returns 63. This will fail in verification when we're comparing with our software state. As per BRCM, this is fixed in 6.5.22, so adding this hack until we upgrade

Differential Revision: D27898365

fbshipit-source-id: 1d21f7fac9bca4fdb8685921994a7b6847467d95