Introduce MultiPimPlatformSystemContainer::getPimType()
Summary:
In this diff, we have all Minipack family system container to implement getPimType()

Currently, we have MinipackPimController::getPimType() to decide the pim type.
However, MinipackPimController is a private unique_ptr of MinipackPimContainer.
And MinipackSystemContainer controls the MinipackPimContainer.

This design might not work if we need to use the pim type to create the pim container, like ElbertSystemContainer.

Therefore, I moved the getPimType() from the controller to the system container.

Differential Revision: D27917651

fbshipit-source-id: 80d26ca80f21fcea26f01459c82acb4910ca8448