add hw asic api for shim header offset
Summary:
Adding an api to get the shim header offset based on platform.

For bcm, the shim header will carry the logical port ID already and offset will be 0.
For Tajo, first few IDs are internally allocated and cannot be used. Hence the shim have an additional offset added to the shim header.

Reviewed By: jasmeetbagga

Differential Revision: D27935436

fbshipit-source-id: 48f1c37fed783236e9c67efa5e9bf528630c315b