api to get system port id based on platform
Summary:
As titled, adding getSystemPortId to return the system port id computed based on the platform

```
Eg:
Logical port id: 1
For Bcm, system port will be set to std::nullopt and will not be configured.
For Tajo sdk 1.38, system port will be set to std::nullopt and will not be configured.
For Tajo sdk 1.40, system port will be set to 1001 (offset of 1000 + port id) will configured.
```

Subsequent diff uses the API.

Reviewed By: jasmeetbagga

Differential Revision: D27980704

fbshipit-source-id: 0a4d02abd8fccdf1e246cb9f93542fae6d067625