Adding Port and profile id to every Phy manager function to pass it to SaiManager
Summary: The SaiPortManager needs the SwPortId and PortProfileId. It will internally find out the PhyConfig based on that. So adding port and profile id information in all PhyManager function for this purpose. This function will currently keep the PhyPortConfig argument also because for 3 platforms the code is still running in agent side

Reviewed By: jasmeetbagga

Differential Revision: D27615832

fbshipit-source-id: 90deaf93562de89d3a5a78565c9bf87ac2968856