change from manually looping server evb to calling serve()
Summary:
This changes how we drive the event base for our ThriftServer
in the fboss agent. Reason is that there is some default logic in
serve that sets up listening for certificate changes that we would
like to pick up. This was moved out of setup and in to serve in
D25774004.

Internally, serve calls setup(), sets up a SCOPE_EXIT for cleanup,
enables ssl config change callbacks and then loops the evb, which is
effectively what we were doing manually before.

Reviewed By: jasmeetbagga

Differential Revision: D28084039

fbshipit-source-id: caf90b83e9b224688f3956742cc3db1ae7a057a3