Only reset ingress acl if SWITCH_ATTR_INGRESS_ACL is programmed
Summary:
I noticed right now xphy hw_test will crash during shutdown()
because we always resetIngressAcl() any way but IngressAcl is not supported in phy sai_switch.

Technically I could use HwAsic::Feature::SWITCH_ATTR_INGRESS_ACL to gate SaiSwitchManager::resetIngressAcl()
But because this function is called in SaiManagerTable::~SaiManagerTable() destructor, therefore we can't call pure virtual function HwAsic::isSupported()

To work around that, I decided to read from hardware to see whether we can get the programmed IngressAcl value back.
And if there's a not null value, then we program it to null.
Or if we try to get this optional value, in unsupported case, it will return the default value(null).
So that we won't call the set function again.

Reviewed By: jasmeetbagga

Differential Revision: D28166547

fbshipit-source-id: 7a717b6ddb6ce34bd31e6e37cde2e69292d9f9c6