try to reprogram xphy on port flap
Summary:
Right now we won't reprogram the xphy unless the speed/enabled state/prbs settings change. This has caused us to not reprogram the xphy when 200G modules are swapped in place of 100G ones, and later caused port flaps when the agent was restarted.

This change will make us check if the xphy needs to be reprogrammed any time the port state changes so we don't miss these swaps.

This is a hacky fix until we support some way of the qsfp service alerting the wedge agent about module swaps

Reviewed By: rajank7

Differential Revision: D28034867

fbshipit-source-id: d0ada826f22966f4b570d040ebf3b951c99cba66